+++ src/components/layout/layout/Layout.js
@@ -1,5 +1,5 @@
-import FooterCom from "../footer/footer";
-import Navbar from "../navbar/navbar";
+import FooterCom from "../footer/changeFooter";
+import Navbar from "../navbar/changeNavbar";
 import "./Layout.css";
 import { useLocation } from "react-router-dom";
