+++ src/pages/LostPets.js
@@ -3,9 +3,9 @@ import { useLocation, useNavigate } from 'react-router-dom';
 import PetCard from '../components/cards/PetCard/PetCard.jsx';
 import pets from "../databases/petDatabase.js";
 import './LostPets.css';
-import Pagination from '../components/generic/pagination/pagination.js';
-import Filter from '../components/filters/filter/filter.js';
-import FilterContainer from '../components/filters/filterContainer/filterContainer.js';
+import Pagination from '../components/generic/pagination/changePagination.js';
+import Filter from '../components/filters/filter/changeFilter.js';
+import FilterContainer from '../components/filters/filterContainer/changefilterContainer.js';
 import Layout from '../components/layout/layout/Layout.js';
 import ButtonComponent from '../components/generic/button/ButtonComponent.js';
 import { clearFilters } from '../utils/clearFilters.js';
